requirements.txt
attr==0.3.2
brotli==1.1.0
ConfigParser==7.0.0
cryptography==43.0.0
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
HTMLParser==0.0.2
importlib_metadata==7.1.0
ipython==8.12.3
ipywidgets==7.8.2
Jinja2==3.1.4
jnius==1.1.0
keyring==25.3.0
lockfile==0.12.2
lxml==5.3.0
matplotlib==3.9.2
numpy==2.0.1
ordereddict==1.1
Pillow==10.4.0
protobuf==5.27.3
pyOpenSSL==24.2.1
pyOpenSSL==24.2.1
railroad==0.5.0
Sphinx==8.0.2
thread==2.0.5
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
xmlrpclib==1.0.1
zipp==3.19.1
